move IComponentBehaviour
Update import
