+++ packages/react-vapor/docs/src/components/ComponentsInterface.tsx
@@ -6,11 +6,6 @@ export interface IComponent {
     component: ExampleComponent;
 }
 
-export interface IComponentBehaviour {
-    isLoading?: boolean;
-    disabled?: boolean;
-}
-
 export type ExampleComponent = React.ComponentType & {
     description?: string;
     firstTabLabel?: string;
